+++ packages/server/core/src/types/plugins/base.ts
@@ -5,7 +5,6 @@ import type {
   Logger,
   Metrics,
   MiddlewareContext,
-  Monitors,
   Reporter,
   ServerRoute,
 } from '@modern-js/types';
@@ -15,24 +14,6 @@ import type { Render } from '../render';
 import type { ServerPlugin } from './plugin';
 
 export type { FileChangeEvent, ResetEvent } from '@modern-js/plugin';
-export type FallbackReason = 'error' | 'header' | 'query' | `header,${string}`;
-
-export type FallbackContext = {
-  request: Request;
-  monitors?: Monitors;
-};
-
-export type FallbackInput = {
-  reason: FallbackReason;
-  error: unknown;
-  context?: FallbackContext;
-};
-
-export type OnFallback = (
-  reason: FallbackReason,
-  error?: unknown,
-  context?: FallbackContext,
-) => Promise<void>;
 
 export type APIServerStartInput = {
   pwd: string;
@@ -74,7 +55,6 @@ export interface GetRenderHandlerOptions {
   pwd: string;
   routes: ServerRoute[];
   config: UserConfig;
-  onFallback?: OnFallback;
   cacheConfig?: CacheConfig;
   staticGenerate?: boolean;
   metaName?: string;