+++ packages/server/core/src/types/plugins/index.ts
@@ -2,12 +2,8 @@ export * from './plugin';
 export type {
   ServerConfig,
   CacheConfig,
-  OnFallback,
-  FallbackReason,
-  FallbackContext,
   GetRenderHandlerOptions,
   FileChangeEvent,
-  FallbackInput,
   WebServerStartInput,
   APIServerStartInput,
   ServerMiddleware,