+++ tests/integration/server-esm-tsx/src/modern-app-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types='@modern-js/app-tools/types' />